feat(cli): optimize zip output name
