+++ packages/manifest/src/index.ts
@@ -22,7 +22,7 @@ import type {
   WebExtensionManifest,
 } from './types.js';
 
-export { setTargetEnv, getEntryFileVariants, defaultExtensionTarget } from './common.js';
+export { setTargetEnv, getEntryFileVariants, defaultExtensionTarget, readManifestFile } from './common.js';
 
 export * from './types.js';
